% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityJoin.R
\name{edit_dist1_example}
\alias{edit_dist1_example}
\title{Generate Example Strings with Edit Distance 1}
\usage{
edit_dist1_example(avg_len = 25, num_strings = 5000)
}
\arguments{
\item{avg_len}{Average length of the strings.}

\item{num_strings}{Number of strings to generate.}
}
\value{
A character vector of generated strings.
}
\description{
This function generates a random list of \verb{num_strings = 5n} strings
such that each of \code{n} strings has one duplicate, one string with a deleted letter,
one string with an inserted letter, and one string with a substituted letter.
}
\seealso{
\code{\link[=similarityJoin]{similarityJoin}}
}
