\name{PatternJoin-package}
\alias{PatternJoin-package}
\alias{PatternJoin}
\docType{package}
\title{
  String Similarity Joins for Hamming and Levenshtein Distances
}
\description{
  This project is a tool for words edit similarity joins under small (< 3) edit distance constraints. 
  It works for Levenshtein distance and Hamming (with allowed insertions/deletions to the end) distance. 
  The software is specifically optimized for edit similarity joins of amino-acid sequences from TCR/BCR profiles, 
  where the number of words is relatively large (10^5-10^6) and the average length of words is relatively small (~20).
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Daniil Matveev.

Maintainer: Daniil Matveev <dmatveev@sfsu.edu>
}
\references{
  paper in progress
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
