% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilarityJoin.R
\name{buildAdjacencyMatrix}
\alias{buildAdjacencyMatrix}
\title{Build Adjacency Matrix}
\usage{
buildAdjacencyMatrix(
  strings,
  cutoff,
  metric = "Levenshtein",
  method = "partition_pattern",
  drop_deg_one = FALSE
)
}
\arguments{
\item{strings}{Input vector of strings.}

\item{cutoff}{Cutoff: 0,1,2.}

\item{metric}{Edit distance type: "Hamming", "Levenshtein".}

\item{method}{Method: "partition_pattern", "pattern", "semi_pattern".}

\item{drop_deg_one}{Drop isolated strings: TRUE, FALSE.}
}
\value{
A sparse adjacency matrix.
}
\description{
Build Adjacency Matrix
}
\examples{
strings <- c("AA", "AB", "CB", "DD")
cutoff <- 1
metric <- "Hamming"
method <- "partition_pattern"
drop_deg_one <- TRUE
adj_matrix <- buildAdjacencyMatrix(strings, cutoff, metric, method, drop_deg_one)
}
